+++ src/revit-plugin/RevitMCPCommandSet/Commands/View/ActivateViewEventHandler.cs
@@ -2,12 +2,12 @@ using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using Autodesk.Revit.UI.Events;
 using RevitMCPSDK.API.Interfaces;
-using RevitMCPViewCommandSet.Extensions;
+using RevitMCPCommandSet.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Threading;
 
-namespace RevitMCPViewCommandSet.Commands.View
+namespace RevitMCPCommandSet.Commands.View
 {
     public class ActivateViewEventHandler : IExternalEventHandler, IWaitableExternalEventHandler
     {